@@ -3,6 +3,8 @@ package main
 import (
 	"fmt"
 	"os"
+
+	"github.com/MrJeffLarry/redmine-cli/internal/cmd"
 )
 
 var version = "0.0.0-dev"
@@ -10,7 +12,8 @@ var commit = ""
 var date = ""
 
 func main() {
-	if err := CmdInit(version, commit, date); err != nil {
+	r := cmd.CmdInit(version, commit, date)
+	if err := r.Cmd.Execute(); err != nil {
 		fmt.Fprintf(os.Stderr, "%s\n", err)
 		os.Exit(1)
 	}
